sentiment-ai-poc/requirements.txt
# Core ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
xgboost>=2.0.0

# Deep Learning
# TensorFlow with GPU support
tensorflow>=2.15.0
# PyTorch with CUDA for Stanza GPU acceleration
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# NLP
nltk>=3.8.0
gensim>=4.3.0
stanza>=1.7.0

# Data
datasets>=2.14.0
huggingface-hub>=0.17.0
pyarrow>=14.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
joblib>=1.3.0

# Hyperparameter Optimization
optuna>=3.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
